requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
distro==1.9.0
Django==3.2.25
django-allauth==0.63.3
django-ratelimit==4.1.0
django-rest-auth==0.9.5
djangorestframework==3.15.1
docker==4.4.4
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.7
jsonschema==3.2.0
paramiko==3.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==5.3.1
ratelimit==2.2.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
texttable==1.7.0
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==0.59.0
